Removed old deployment files
