--- protocol/deploys/mainnet/temple-1m-vault-verify/temple-1m-vault-a.js
@@ -1,12 +1,0 @@
-module.exports = [
-    '1m-core',
-    '1m-core-a',
-    '0x470ebf5f030ed85fc1ed4c2d36b9dd02e77cf1b7', // temple erc20 mainnet
-    '0xc3940f86a16f54a6d74e200616eb7309e31ac004', // temple exposure
-    '0xec3c1abdab15ebc069ec5e320eaacf716edfc011', // vaulted temple
-    2419200,
-    604800,
-    {p:1,q:1},
-    '0x8a17403b929ed1b6b50ea880d9c93068a5105d4c',
-    1654980754 
-]
