--- protocol/deploys/rinkeby/30-min-vault-verify/tst-30min-f.js
@@ -1,10 +1,0 @@
-module.exports = [
-    'tst-30min',
-    'tst-30min-f',
-    '0x359655dcB8A32479680Af81Eb38eA3Bb2B42Af54',
-    1800,
-    300,
-    {p:1,q:1},
-    '0x8dFEeCCacda824450a4e912454d6ba0E6aF24c1a',
-    1654049670
-]
